# src/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Express API with Swagger",
    "description": "This is a simple CRUD API application made with Express and documented with Swagger",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic",
      "in": "header",
      "name": "Authorization",
      "value": "Basic <user:password>"
    },
    "apiKeyAuth": {
      "type": "apiKey",
      "name": "X-API-KEY",
      "in": "header"
    },
    "bearerAuth": {
      "type": "http",
      "name": "Authorization",
      "in": "header",
      "scheme": "bearer"
    }
  },
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/id/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/score/redis/{key}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/score/sqlite/{Id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    }
  },
  "definitions": {
    "Score": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "playerName": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "stage": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                }
              }
            },
            "score": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                }
              }
            },
            "playTime": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                }
              }
            },
            "dateTime": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "timestamp": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "timeZone": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}